+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-06-16 21:06:57
-// Commit: feat(auth): add bi-directional event bus handler for auth
+// Updated at: 2026-06-16 21:13:05
+// Commit: chore(auth): update dependencies and cleanup unused imports in auth
 
 class authComponentState {
-  final String commitRef = '2049';
-  final String lastMessage = 'feat(auth): add bi-directional event bus handler for auth';
-  final DateTime timestamp = DateTime.parse('2026-06-16T21:06:57+05:00');
+  final String commitRef = '2051';
+  final String lastMessage = 'chore(auth): update dependencies and cleanup unused imports in auth';
+  final DateTime timestamp = DateTime.parse('2026-06-16T21:13:05+05:00');
 }